fix numpy dataset bug

+++ myutils/datasets.py
@@ -11,10 +11,10 @@ from torchvision import transforms, utils
 class NumpyDataset(Dataset):
     def __init__(self, root_dir, train=False):
         if train:
-            self.data = torch.tensor(np.load(os.path.join(root_dir,'train_data.npy')))
+            self.data = torch.tensor(np.load(os.path.join(root_dir,'train_data.npy')), dtype=torch.float32)
             self.labels = torch.tensor(np.load(os.path.join(root_dir, 'train_label.npy')))
         else:
-            self.data = torch.tensor(np.load(os.path.join(root_dir,'val_data.npy')))
+            self.data = torch.tensor(np.load(os.path.join(root_dir,'val_data.npy')), dtype=torch.float32)
             self.labels = torch.tensor(np.load(os.path.join(root_dir, 'val_label.npy')))
     
     def __len__(self):
